+++ R/namespace.R
@@ -1,2 +1,3 @@
 #' @import checkr hms
+#' @importFrom chk chk_range
 NULL