+++ NEWS.md
@@ -1,3 +1,4 @@
+- `dtt_doy_to_date()` now accepts whole numeric (as well as integer)
 - Replaced deprecated `hms::as.hms()` with `hms::as_hms()`
 
 # dttr 0.0.1